Improve QDBusInterface error-handling
As it turns out, it is possible for interface.isValid() to return False
but have a QDBusError of type 0 (which, according to the Qt5 docs,
indicates that the "QDBusError is invalid (i.e., the call succeeded)").
Accordingly, an exception should raise only for a non-zero error type
(and RuntimeError should be used instead since any number of things
could have gone wrong).