Fix MainWindow 'populate' method
This change updates the MainWindow.populate method to behave in a more
"additive" manner: calling it will no longer clear/re-create the
associated QComboBox items and folders/history views widgets, thereby
fixing the bug which required users to re-start the application in the
event of joining a new grid via the Welcome dialog due to disconnected
Qt Signals/Slots. In addition, adding a new grid will now select it
immediately in the combobox, making it more obvious that the new grid
has been joined and facilitating adding new folders afterward.