Add support for specifying a "logo" via config.txt
This adds support for specifying a 'logo_icon' application setting via
'config.txt'. If 'logo_icon' is set to a resource file, that file will
be loaded/displayed on the initial "Welcome" screen in place of the
'tray_icon'. In addition, if a 'logo_icon' is set, hide the application
description/slogan (since "logos", by definition, contain a "mark" and a
"type" -- i.e., an image and text combination -- and so we probably
don't want/need the description/slogan text)