Use a reproducible build directories, when passing `--require-hashes` to pip.
When `--require-hashes` is passed to pip, it will not use locally build wheels
in its cache, for packages that are only available as sdists. This causes it to
always rebuild those packages from source.

Since extension modules (like pyrsistent) embeded the build directory, which is
randomly generated, while the set of *packages* is reproducible, the installed
modules aren't.

To work around this, we add a wrapper around pip, that forces it to use
reproducible temporary directory names.