Update Deferred type wrapper for Twisted 21.7.0
Twisted 21.7.0 adds type hinting for twisted.internet.defer, changing
the behavior of this supertype. Unfortunately, it appears that mypy
still expects Generators for any functions decorated by inlineCallbacks
(so a bare Deferreds can't be used as a return type annotations in such
cases -- and thus this wrapper/workaround is still needed).