Remove the Nix expressions for packaging GridSync with Nix
Due to dependency changes in GridSync these expressions no longer successfully
build a package.

In the future, we should try again with a tool like mach-nix which can read
the same dependency information as the Python packaging tools.  This reduced
duplication will hopefully make it easier to keep the Nix expressions up to
date and working.

Additionally, we should have CI for the Nix packaging so that developers are
reminded when they need to do something extra to keep the Nix packages working.