configure the warnings module, via pytest, to ignore ResourceWarning
CPython makes ResourceWarning off by default but pytest turns it on by default
so we supply pytest configuration to turn it off again.

This should mitigate the CPython bug because ignored warnings are detected
before the warnings module does any imports (and they are handled by exiting
early so the imports never happen).