[wip] sketch/spike a way we might manage log privacy with a logging Filter
This also includes pieces to let a user choose their privacy level on the
command line or to have a default privacy setting derived from grid
configuration.

This is intended to support cases like "preserve privacy by default on the
production grid", "reveal private information by default on the staging grid",
and "reveal private information to a developer interacting with the production grid".